+++ src/main/java/edu/hm/cs/pblv/group3/entities/CocktailIngredient.java
@@ -1,4 +1,4 @@
-package edu.hm.cs.pblv.group3.entities;
+package main.java.group3.entities;
 
 public enum CocktailIngredient {
